+++ pysimavr/swig/utils.i
